Impl uint32 steps and sps

+++ src/rp2350/shared.h
@@ -58,8 +58,8 @@
 struct Segment {
     uint8_t  numMotors;           // How many motors are in this specific move
     uint8_t  nodeId[MAX_MOTORS];  // Node IDs for this move (1 to 4)
-    uint16_t steps[MAX_MOTORS];
-    uint16_t sps[MAX_MOTORS];
+    uint32_t steps[MAX_MOTORS];
+    uint32_t sps[MAX_MOTORS];
     bool     cw[MAX_MOTORS];
 };
